requirements.txt
pandas==1.5.3
python-dotenv==1.0.1
tqdm==4.66.4
icecream==2.1.3

langchain==0.2.7
langchain-core==0.2.18
langchain-openai==0.1.16